+++ src/main/resources/logback.xml
@@ -17,7 +17,13 @@
             </pattern>
         </encoder>
     </appender>
+    <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>%d{ISO8601} [%t]%X{space}%X{requestId}%X{space}%X{userEmail} %p %c - %m%n</pattern>
+        </encoder>
+    </appender>
     <root level="DEBUG">
         <appender-ref ref="FILE"/>
+        <appender-ref ref="CONSOLE"/>
     </root>
 </configuration>
